Refactoring dashboard route to user_devices route. Splitting DAO file into multiple files with each DAO class.
