+++ src/controller/GraphController.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from datetime import datetime
 
-from ..database.dao import DeviceOccupancyDao
+from ..dao import DeviceOccupancyDao
 
 
 class GraphController:
